Price using BigDecimal instead of plain double

+++ core/src/main/java/live/dobbie/core/misc/Price.java
@@ -4,10 +4,12 @@ import live.dobbie.core.context.primitive.converter.ConvertableToString;
 import lombok.NonNull;
 import lombok.Value;
 
+import java.math.BigDecimal;
+
 @Value
 @ConvertableToString
 public class Price {
-    double amount;
+    BigDecimal amount;
     @NonNull Currency currency;
 
     @Override
